chore: 커피샵의 메뉴를 인스턴스 생성 시 구성.
