+++ CodeStarterCamp_Week3/Coffee.swift
@@ -1,0 +1,12 @@
+//
+//  Coffee.swift
+//  CodeStarterCamp_Week3
+//
+//  Created by jeseoyoung on 2024/10/04.
+//
+
+import Foundation
+
+enum Coffee {
+    case 라떼, 아메리카노, 콜드브루, 디카페인
+}
