+++ CodeStarterCamp_Week3/Person.swift
@@ -1,0 +1,26 @@
+//
+//  Person.swift
+//  CodeStarterCamp_Week3
+//
+//  Created by jeseoyoung on 2024/10/04.
+//
+
+import Foundation
+
+struct Person {
+    var appearance: String
+    var personality: String
+    var budget: Int
+    var height: Double
+    
+    init(appearance: String, personality: String, budget: Int, height: Double) {
+        self.appearance = appearance
+        self.personality = personality
+        self.budget = budget
+        self.height = height
+    }
+    
+    func buyCoffee() {
+    
+    }
+}
